rbac: fix permission decorator for global permissions
